old/requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
decouple==0.0.7
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-webpack-loader==0.6.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
py-mini-racer==0.6.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
whitenoise==6.4.0
